requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.7.0
openai==1.58.1
python-telegram-bot==21.9
requests~=2.32.3
tenacity==8.3.0
wolframalpha~=5.1.3
duckduckgo_search==7.1.1
spotipy~=2.24.0
pytube~=15.0.0
gtts~=2.5.4
whois~=0.9.27
Pillow~=11.0.0
opencv-python~=4.9.0
numpy~=1.26.0
uvicorn
yfinance
ta
goose3~=3.1.12
playwright==1.42.0
